Add files via upload
In this code, you'll be prompted to enter the temperature for each day of the month (assuming a maximum of 31 days). The temperatures will be stored in a list, and then the average temperature in a day and at the end of the month will be calculated by dividing the sum of all temperatures by the corresponding number of days. Finally, the code will print the average temperatures.